framwork 2.0#201505024-1
******************************************************
版本号：2.0#201505024-1

主要变更：

1、service::Runable析构时自动停止运行中服务
2、transport::Whisper增加获取监听ip方法transport::Whisper::ip()
3、重构service::TimerService，改为无锁设计，性能更高，并发更强，触发更精准，timer
id类型变更为uint32_t，0代表无效定时器
4、事务内定时器同步调整为uint32_t

兼容性说明：
1、定时器id类型调整为uint32_t，0代表无效定时器

+++ framework/include/transaction/base_task.hpp
@@ -72,7 +72,7 @@ namespace zhicloud{
                     return;
                 }
                 if((AppMessage::message_type::EVENT == msg.type)&&((uint32_t)EventEnum::timeout == msg.id)){
-                    if(msg.sequence != (uint32_t)session.getTimerID()){
+                    if(msg.sequence != session.getTimerID()){
                         logger->warn(boost::format("[%08X]ignore timeout event with timer id %d(current %d)")
                                      %session.getSessionID() %msg.sequence %session.getTimerID());
                         return;
